Improved task 2477

+++ src/main/java/g2401_2500/s2477_minimum_fuel_cost_to_report_to_the_capital/Solution.java
@@ -24,7 +24,7 @@ public class Solution {
     }
 
     private long solve(ArrayList<ArrayList<Integer>> adj, int seats, int src, int parent) {
-        int people = 1;
+        long people = 1L;
         for (int i : adj.get(src)) {
             if (i != parent) {
                 people += solve(adj, seats, i, src);
